--- golang/boot-go-http-client/01-why-http/go.mod
@@ -1,3 +1,0 @@
-module 01-why-http
-
-go 1.22.2
