+++ golang/boot-go-http-client/chap01-why-http/get-issue.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	"01-why-http/localpackage/myhttp"
+	"chap01-why-http/localpackage/myhttp"
 )
 
 const issueURL = "https://api.boot.dev/v1/courses_rest_api/learn-http/issues"
