+++ golang/boot-go-http-client/chap01-why-http/get-user.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	"01-why-http/localpackage/myhttp"
+	"chap01-why-http/localpackage/myhttp"
 )
 
 const usersUrl = "https://api.boot.dev/v1/courses_rest_api/learn-http/users"
