+++ golang/boot-go-http-client/chap01-why-http/go.mod
@@ -1,0 +1,3 @@
+module chap01-why-http
+
+go 1.22.2
