+++ golang/boot-go-http-client/chap01-why-http/localpackage/myhttp/myhttp.go
